+++ xarray_enmap/xarray_enmap.py
@@ -33,8 +33,11 @@ VAR_MAP = dict(
     haze="QL_QUALITY_HAZE",
     snow="QL_QUALITY_SNOW",
     testflags="QL_QUALITY_TESTFLAGS",
-    # We omit the quicklook files QL_SWIR and QL_VNIR.
+    swirquicklook="QL_SWIR",
+    vnirquicklook="QL_VNIR"
 )
+
+QUICKLOOK_VAR_NAMES = ["swirquicklook", "vnirquicklook"]
 
 
 class EnmapEntrypoint(xr.backends.BackendEntrypoint):
@@ -116,6 +119,10 @@ def read_dataset_from_inner_directory(
         name: rioxarray.open_rasterio(filename).squeeze()
         for name, filename in find_datafiles(data_path).items()
     }
+    for quicklook_name in QUICKLOOK_VAR_NAMES:
+        if quicklook_name in arrays.keys():
+            ql = arrays.get(quicklook_name)
+            arrays[quicklook_name] = ql.rename({"band": "quicklookband"})
     if "reflectance" in arrays.keys() and scale_reflectance:
         reflectance = arrays.get("reflectance")
         fill_value = reflectance.attrs.get("_FillValue", -32768)